44.站内消息通知（Part2）/requirements.txt
Django==2.0
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-model-utils==3.1.2
django-notifications-hq==1.5.0
jsonfield==2.0.2
mysqlclient==1.3.13
Pillow==5.0.0
pytz==2017.3
